add specular glass material & matte material (with bug)

+++ WRender/Shaders/RayTracing/BxDF/BSDFCommon.hlsl
@@ -66,37 +66,9 @@ inline float Sin2Phi(float3 w)
     return SinPhi(w) * SinPhi(w);
 }
 
-//float FrDielectric(float cosThetaI, float etaI, float etaT)
-//{
-//    cosThetaI = clamp(cosThetaI, -1, 1);
-//    // Potentially swap indices of refraction
-//    bool entering = cosThetaI > 0.f;
-//    if (!entering)
-//    {
-//        float tmp = etaI;
-//        etaI = etaT;
-//        etaT = tmp;
-//        cosThetaI = abs(cosThetaI);
-//    }
-
-//    // Compute _cosThetaT_ using Snell's law
-//    float sinThetaI = sqrt(max((float) 0, 1 - cosThetaI * cosThetaI));
-//    float sinThetaT = etaI / etaT * sinThetaI;
-
-//    // Handle total internal reflection
-//    if (sinThetaT >= 1)
-//        return 1;
-//    float cosThetaT = sqrt(max((float) 0, 1 - sinThetaT * sinThetaT));
-//    float Rparl = ((etaT * cosThetaI) - (etaI * cosThetaT)) /
-//                  ((etaT * cosThetaI) + (etaI * cosThetaT));
-//    float Rperp = ((etaI * cosThetaI) - (etaT * cosThetaT)) /
-//                  ((etaI * cosThetaI) + (etaT * cosThetaT));
-//    return (Rparl * Rparl + Rperp * Rperp) / 2;
-//}
-
 bool MatchesFlags(BxDFType type, BxDFType t)
 {
-    return ((uint) type & (uint) t) == (uint) type;
+    return (type & t) == type;
 }
 
 #endif
